add: adiciona middleware de verificação de signUp
